change: Reorganize integ test files for workflow
